+++ image_cropper_for_web/CHANGELOG.md
@@ -1,3 +1,16 @@
+
+## [1.0.2] - 2022-08-09
+
+* correct `WebSettings` model
+
+## [1.0.1] - 2022-08-09
+
+* correct default values of `WebUiSettings`
+
+## [1.0.0] - 2022-08-09
+
+* **BREAKING CHANGE**: move all setting models to platform interface packages
+
 ## [0.0.4] - 2022-05-17
 
 * correct importing JS library