+++ image_cropper_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [3.0.2] - 2022-08-09
+
+* correct `WebSettings` model
+
+## [3.0.1] - 2022-08-09
+
+* correct default values of `WebUiSettings`
+
+## [3.0.0] - 2022-08-09
+
+* **BREAKING CHANGE**: move all setting models to platform interface packages
+
 ## [2.0.0-beta.1] - 2022-03-27
 
 * **BREAKING CHANGE**: update result models to support web, replace `File` by `CroppedFile`